refactor(api): :art: create new repository for feature flag and version tag relation
